Adding states and statemachine
